requirements/local.txt
# General requirements
docutils==0.12
ipdb==0.10.0
Werkzeug>=0.9.6

# Django requirements
django-debug-toolbar>=1.8
django-pdb==0.5.1
